Enable Server Name Indication Support - Initial / Reference commit
This is on a different branch than the [Python/Ruby SNI mods][1] just to make
testing easier for myself. I.e. I had code that worked with the 2.3.3
driver where as the recent [Jackson changes][2] broke that. However, the
underlying SNI changes will work with the Next branch.

This includes some commented out printlns that ultimately need removing,
but they were for my debugging and I don't want to forget that.

Enabling SNI itself is easy, the tricky bit for me was modifying a test
app to work with it. You have to use the SSLContext option similar to
[here][3]. For testing I've just used a null sslContext:

	private static final String DEFAULT_SSL_PROTOCOL = "TLSv1.2";

	...

	Connection conn = null;
	Connection.Builder builder = r.connection().
	    hostname(host).
	    port(port).
	    authKey(authKey);

	final SSLContext sslContext = SSLContext.getInstance(DEFAULT_SSL_PROTOCOL);
	sslContext.init(null, null, null);
	conn = builder.sslContext(sslContext).connect();

[1]: https://github.com/atomicules/rethinkdb/commits/sni
[2]: https://github.com/rethinkdb/rethinkdb/pull/6157
[3]: https://github.com/pires/rethinkdb-ssl-test/blob/master/src/main/java/com/github/pires/App.java#L70